+++ app/assets/CSS/style.css
@@ -18,6 +18,13 @@ body{background-color: var(--c-bg);border-radius: 5px;padding: 5px;width: 500px;
     padding-top: 5px;
     padding-left: 5px;
 }
+.pick{
+    position: absolute;
+    margin-top: -85%;
+    margin-left: 18%;
+    text-align: center;
+    pointer-events: none;
+}
 p{margin:0}h1{margin:5px}
 .titre{
     color:whitesmoke;
