+++ app/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="fr">
     <head>
-        <title>ColorPicker</title>
+        <title>Color Picker</title>
         <META NAME="Author" CONTENT="DevilishDante">
         <META NAME="Description" CONTENT="Color Picker by DevilishDante">
         <META NAME="viewport" CONTENT="width=device-width, initial-scale=1.0">
@@ -10,8 +10,9 @@
         <META HTTP-EQUIV="content-type" CONTENT="text/html; charset=UTF-8">
         <!--CSS BASICS-->
         <link rel="stylesheet" href="./assets/CSS/style.css" type="text/css">
+        <!-- ADDONS -->
         <link rel="stylesheet" href="./assets/CSS/tooltip.css" type="text/css">
-        <script src="./assets/js/core.js" defer></script>
+        <script type="module" src="./assets/js/core.js" defer></script>
     </head>
     <body>
         <img src="./assets/icons/icon-48.png" class="logo" alt="logo">
@@ -19,7 +20,13 @@
         <div class="window">
             <div class="flex-container">
                 <div class="flex-item center color-main">
-                    <tooltip><input type="color" id="pick-color" value="#D66853">
+                    <tooltip>
+                        <input type="color" id="pick-color" value="#D66853">
+                        <div class="pick">
+                            <svg id="pick-icon" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" fill="#000000" height="3rem" width="3.5rem" version="1.1" id="Icons" viewBox="0 0 32 32" xml:space="preserve">
+                                <path d="M27.7,3.3c-1.5-1.5-3.9-1.5-5.4,0L17,8.6l-1.3-1.3c-0.4-0.4-1-0.4-1.4,0s-0.4,1,0,1.4l1.3,1.3L5,20.6  c-0.6,0.6-1,1.4-1.1,2.3C3.3,23.4,3,24.2,3,25c0,1.7,1.3,3,3,3c0.8,0,1.6-0.3,2.2-0.9C9,27,9.8,26.6,10.4,26L21,15.4l1.3,1.3  c0.2,0.2,0.5,0.3,0.7,0.3s0.5-0.1,0.7-0.3c0.4-0.4,0.4-1,0-1.4L22.4,14l5.3-5.3C29.2,7.2,29.2,4.8,27.7,3.3z M9,24.6  c-0.4,0.4-0.8,0.6-1.3,0.5c-0.4,0-0.7,0.2-0.9,0.5C6.7,25.8,6.3,26,6,26c-0.6,0-1-0.4-1-1c0-0.3,0.2-0.7,0.5-0.8  c0.3-0.2,0.5-0.5,0.5-0.9c0-0.5,0.2-1,0.5-1.3L17,11.4l2.6,2.6L9,24.6z"/>
+                            </svg>
+                        </div>
                         <content class="t-content t-right">Click &<br>Pick !</content>
                     </tooltip>
                     <input type="button" id="save-color" value="save">
